requirement.txt
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.67.0
h11==0.12.0
importlib-metadata==4.6.1
numpy==1.21.1
Pillow==8.3.1
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
uvicorn==0.14.0
zipp==3.5.0
